extract curl
