+++ packages/insomnia-app/app/global.d.ts
@@ -24,7 +24,7 @@ interface Window {
   __REDUX_DEVTOOLS_EXTENSION_COMPOSE__: Function;
   main: {
     restart: () => void;
-    authorizeUserInWindow: (options: { url: string; urlSuccessRegex?: RegExp; urlFailureRegex?: RegExp; sessionId: string }) => Promise<string>;
+    authorizeUserInWindow: (options: { url: string; urlSuccessRegex?: RegExp; urlFailureRegex?: RegExp }) => Promise<string>;
     setMenuBarVisibility: (visible: boolean) => void;
     installPlugin: (url: string) => void;
   };
