+++ packages/insomnia-app/app/main/grpc-ipc-main.ts
@@ -1,6 +1,6 @@
 import { ipcMain } from 'electron';
 
-import { GrpcRequestEventEnum } from '../common/grpc-events';
+import { GrpcRequestEventEnum } from '../common/ipc-events';
 import * as grpc from '../network/grpc';
 import { GrpcIpcRequestParams } from '../network/grpc/prepare';
 import { ResponseCallbacks } from '../network/grpc/response-callbacks';
