+++ packages/insomnia-app/app/ui/components/panes/response-pane.tsx
@@ -1,6 +1,6 @@
 import { autoBindMethodsForReact } from 'class-autobind-decorator';
 import classnames from 'classnames';
-import { clipboard } from 'electron';
+import { clipboard, ipcRenderer } from 'electron';
 import fs from 'fs';
 import { HotKeyRegistry } from 'insomnia-common';
 import { json as jsonPrettify } from 'insomnia-prettify';
@@ -17,7 +17,6 @@ import type { Request } from '../../../models/request';
 import type { RequestVersion } from '../../../models/request-version';
 import type { Response } from '../../../models/response';
 import type { UnitTestResult } from '../../../models/unit-test-result';
-import { cancelRequestById } from '../../../network/network';
 import { Button } from '../base/button';
 import { PreviewModeDropdown } from '../dropdowns/preview-mode-dropdown';
 import { ResponseHistoryDropdown } from '../dropdowns/response-history-dropdown';
@@ -252,7 +251,7 @@ export class ResponsePane extends PureComponent<Props> {
       return (
         <PlaceholderResponsePane hotKeyRegistry={hotKeyRegistry}>
           <ResponseTimer
-            handleCancel={() => cancelRequestById(request._id)}
+            handleCancel={() => ipcRenderer.send('cancelRequestById', request._id)}
             loadStartTime={loadStartTime}
           />
         </PlaceholderResponsePane>
@@ -367,7 +366,7 @@ export class ResponsePane extends PureComponent<Props> {
         </Tabs>
         <ErrorBoundary errorClassName="font-error pad text-center">
           <ResponseTimer
-            handleCancel={() => cancelRequestById(request._id)}
+            handleCancel={() => ipcRenderer.send('cancelRequestById', request._id)}
             loadStartTime={loadStartTime}
           />
         </ErrorBoundary>
