+++ packages/insomnia-app/app/ui/context/grpc/grpc-ipc-renderer.ts
@@ -1,7 +1,7 @@
 import { ipcRenderer } from 'electron';
 
-import type { GrpcRequestEvent } from '../../../common/grpc-events';
-import { GrpcResponseEventEnum } from '../../../common/grpc-events';
+import type { GrpcRequestEvent } from '../../../common/ipc-events';
+import { GrpcResponseEventEnum } from '../../../common/ipc-events';
 import type { GrpcDispatch } from './grpc-actions';
 import { grpcActions } from './grpc-actions';
